+++ include/igraph_bipartite.h
@@ -71,6 +71,14 @@ IGRAPH_EXPORT igraph_error_t igraph_biadjacency(
     igraph_neimode_t mode,
     igraph_bool_t multiple);
 
+IGRAPH_EXPORT igraph_error_t igraph_weighted_biadjacency(
+    igraph_t *graph,
+    igraph_vector_bool_t *types,
+    igraph_vector_t *weights,
+    const igraph_matrix_t *biadjmatrix,
+    igraph_bool_t directed,
+    igraph_neimode_t mode);
+
 IGRAPH_EXPORT igraph_error_t igraph_get_biadjacency(const igraph_t *graph,
                                                     const igraph_vector_bool_t *types,
                                                     const igraph_vector_t *weights,