+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 
  - `igraph_erdos_renyi_game_gnm()` gained a `multiple` Boolean argument to generate Erdős-Rényi graphs with multi-edges
  - `igraph_bipartite_game_gnm()` gained a `multiple` Boolean argument to generate random bipartite graphs with multi-edges
+ - `igraph_weighted_biadjacency()` created a weighted graph from a bipartite adjacency matrix.
 
 ### Changed